Because Nancy middleware is in own project, move middlware tests to Nancy.Tests
