+++ src/Nancy.Owin.Tests/Nancy.Owin.Tests.csproj
@@ -98,9 +98,6 @@
       <Link>Properties\SharedAssemblyInfo.cs</Link>
     </Compile>
     <Compile Include="AppBuilderExtensionsFixture.cs" />
-    <Compile Include="NancyOptionsExtensionsFixture.cs" />
-    <Compile Include="NancyOptionsFixture.cs" />
-    <Compile Include="NancyMiddlewareFixture.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Nancy.Owin\Nancy.Owin.csproj">
